refactor(redis): integrate futures for async handling in Redis backend operations and update dependencies

+++ crates/adapter/src/redis/primitives/set.rs
@@ -188,7 +188,7 @@ impl RedisSet {
     /// Sets the TTL of a set in seconds
     pub fn expire(&self, key: &str, seconds: u64) -> RedisResult<bool> {
         let mut conn = self.acquire_connection()?;
-        let result: i32 = conn.expire(key, seconds as usize)?;
+        let result: i32 = conn.expire(key, seconds as i64)?;
         Ok(result == 1)
     }
